Created and implemented a new global debug variable for printing. Also changed timestamp precision

+++ src/pack-info.hpp
@@ -9,8 +9,10 @@ struct packet_info
     uint32_t dst_ip;
 
     uint32_t ip_octets;
+
     // timestamp relative to the export program
-    int32_t relative_timestamp;
+    // this should be enough precision to even allow for parsing pcap files created at the start of the epoch
+    int64_t relative_timestamp;
     
     uint16_t src_port;
     uint16_t dst_port;
